make the aspect template more readable
